refactor: extract SIWE helpers to services/siwe.go
siweMessage and verifyEthSignature shared between auth_service and
upcoming user_service.LinkWallet. Removes go-ethereum/crypto import
from auth_service.go.

refs #249
